requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.0
Flask-DebugToolbar==0.16.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
mysql-connector-python==9.1.0
numpy==2.2.1
packaging==24.2
pandas==2.2.3
plotly==5.24.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
six==1.17.0
tenacity==9.0.0
tzdata==2024.2
Werkzeug==3.1.3
